--- src/main/java/com/ironknee/Knee2KneelSpring/service/stage/StageDifficulty.java
@@ -1,5 +1,0 @@
-package com.ironknee.Knee2KneelSpring.service.stage;
-
-public enum StageDifficulty {
-    computer_science, korean_literature, english_literature
-}
