+++ src/main/java/com/ironknee/Knee2KneelSpring/service/stage_unused/StageDepartment.java
@@ -1,4 +1,4 @@
-package com.ironknee.Knee2KneelSpring.service.stage;
+package com.ironknee.Knee2KneelSpring.service.stage_unused;
 
 public enum StageDepartment {
     computer_science, korean_literature, english_literature
