+++ src/main/java/com/ironknee/Knee2KneelSpring/service/stage_unused/StageDifficulty.java
@@ -1,0 +1,5 @@
+package com.ironknee.Knee2KneelSpring.service.stage_unused;
+
+public enum StageDifficulty {
+    easy, medium, hard
+}
