+++ src/main/java/com/ironknee/Knee2KneelSpring/service/stage_unused/StageService.java
@@ -1,4 +1,4 @@
-package com.ironknee.Knee2KneelSpring.service.stage;
+package com.ironknee.Knee2KneelSpring.service.stage_unused;
 
 import org.springframework.stereotype.Service;
